Move PostListItemUiState logic out of VM
