+++ WordPress/src/main/java/org/wordpress/android/ui/posts/PostAdapterItem.kt
@@ -1,32 +1,6 @@
 package org.wordpress.android.ui.posts
 
-import org.wordpress.android.fluxc.model.post.PostStatus
 import org.wordpress.android.fluxc.store.UploadStore.UploadError
-import org.wordpress.android.widgets.PostListButtonType
-
-class PostAdapterItem(
-    val data: PostAdapterItemData,
-    val onSelected: () -> Unit,
-    val onButtonClicked: (PostListButtonType) -> Unit
-)
-
-data class PostAdapterItemData(
-    val localPostId: Int,
-    val remotePostId: Long?,
-    val title: String?,
-    val excerpt: String?,
-    val isLocalDraft: Boolean,
-    val date: String,
-    val postStatus: PostStatus,
-    val isLocallyChanged: Boolean,
-    val isConflicted: Boolean,
-    val canShowStats: Boolean,
-    val canPublishPost: Boolean,
-    val canRetryUpload: Boolean,
-    val featuredImageId: Long,
-    val featuredImageUrl: String?,
-    val uploadStatus: PostAdapterItemUploadStatus
-)
 
 data class PostAdapterItemUploadStatus(
     val uploadError: UploadError?,
